+++ src/zoramodulemanager.ts
@@ -35,11 +35,9 @@ export class ZoraModuleManager {
     chainid: number,
     contractAddress?: string
   ) {
-    if (Signer.isSigner(signerOrProvider)) {
-      this.readOnly = false;
-    } else {
-      this.readOnly = true;
-    }
+    Signer.isSigner(signerOrProvider)
+      ? (this.readOnly = false)
+      : (this.readOnly = true);
 
     this.signerOrProvider = signerOrProvider;
     this.chainId = chainid;
@@ -155,6 +153,23 @@ export class ZoraModuleManager {
       r,
       s
     );
+  }
+
+  // registerModule
+  // @note used for testing purposes only
+  // @param {string} moduleAddress Address of the module to register
+  // @returns {Promise<ContractTransaction>} Transaction object
+  public async registerModule(
+    moduleAddress: string
+  ): Promise<ContractTransaction> {
+    try {
+      this.ensureNotReadOnly();
+      validateAndParseAddress(moduleAddress);
+    } catch (err) {
+      return Promise.reject(err);
+    }
+
+    return this.contract.registerModule(moduleAddress);
   }
 
   // Private methods