+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.9] - 2022-02-15
+
+### Added
+
+- Added `RegisterModule` to `ZoraModuleManager` (used for testing)
+- Added WETH utilities to `utils`
+- Improved coverage for both `AsksV11` and `ZoraModuleManager`, write functions now covered/tested
+- Cleanup, removal of unused imports
+
 ## [0.0.8] - 2022-02-14
 
 ### Added
@@ -64,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial pubish of cdk package
 
+[0.0.9]: https://github.com/catalogworks/cdk/packages/1173720?version=0.0.9
 [0.0.8]: https://github.com/catalogworks/cdk/packages/1173720?version=0.0.8
 [0.0.7]: https://github.com/catalogworks/cdk/packages/1173720?version=0.0.7
 [0.0.6]: https://github.com/catalogworks/cdk/packages/1173720?version=0.0.6